renamed Card to Grid, as new search will have a card component
