+++ src/SCSS/App.scss
@@ -9,5 +9,5 @@
 'variables/spacing';
 
 // Components 
-@import 'components/card',
+@import 'components/grid',
 'components/pagination';
